add expect.sequence matcher
